Add separate bean for page new user: AddUserBean which handles all input and logic for this feature. + Rename UserBean to AdminBean: now only responsible for user list and user-detail view!
