+++ src/main/webapp/pages/admin/all-users.xhtml
@@ -16,24 +16,24 @@
                 <h:outputStylesheet name="search-user.css" library="styles" />
             </f:facet>
 
-            #{userBean.fetchAllUsers()}
+            #{adminBean.fetchAllUsers()}
 
             <p:panel id="searchPanel" >
                 <p:panelGrid id="searchPanelGrid" columns="1">
-                    <p:inputText id="searchInput" value="#{userBean.searchString}">
+                    <p:inputText id="searchInput" value="#{adminBean.searchString}">
                         <p:ajax event="keyup" delay="100" update="searchResult"
-                                listener="#{userBean.fetchAllUsers()}" />
+                                listener="#{adminBean.fetchAllUsers()}" />
                     </p:inputText>
                     <p:watermark for="searchInput" value="Nach Benutzern suchen..." id="watermark" />
                 </p:panelGrid>
             </p:panel>
 
             <p:panel id="resultPanel">
-                <p:dataScroller id="searchResult" value="#{userBean.searchResult}" var="user" chunkSize="10">
+                <p:dataScroller id="searchResult" value="#{adminBean.searchResult}" var="user" chunkSize="10">
                     <f:facet name="header">Benutzer</f:facet>
 
                     <h:panelGrid columns="3" style="width:100%" columnClasses="button,editor">
-                        <h:commandButton action="#{userBean.viewUserDetails(user)}" value="Details" id="detailsButton" />
+                        <h:commandButton action="#{adminBean.viewUserDetails(user)}" value="Details" id="detailsButton" />
 
                         <p:outputPanel>
                             <h:panelGrid columns="2" columnClasses="logo,detail">
