+++ src/main/webapp/pages/admin/user-details.xhtml
@@ -19,20 +19,20 @@
             <p:panel id="userDetails" header="#{msgs['admin.details.data']}" styleClass="panelGridCenterAndTop">
                 <h:panelGrid style="width:100%" columns="2" cellpadding="80" columnClasses="description, detail" styleClass="panelGridCenter">
                     <h:outputLabel value="#{msgs['label.name']}: " style="font-weight: bold"/>
-                    <h:outputText value="#{userBean.currentUser.name}" />
+                    <h:outputText value="#{adminBean.currentUser.name}" />
 
                     <h:outputLabel value="#{msgs['label.login_id']}: " style="font-weight: bold"/>
-                    <h:outputText value="#{userBean.currentUser.loginID}" />
+                    <h:outputText value="#{adminBean.currentUser.loginID}" />
 
                     <h:outputLabel value="#{msgs['label.email']}: " style="font-weight: bold"/>
-                    <h:outputText value="#{userBean.currentUser.email}" />
+                    <h:outputText value="#{adminBean.currentUser.email}" />
 
                     <h:outputLabel value="#{msgs['label.company']}: " style="font-weight: bold"/>
-                    <h:outputText value="#{userBean.currentUser.company}" />
+                    <h:outputText value="#{adminBean.currentUser.company}" />
 
                     <h:outputLabel value="#{msgs['label.role']}: " style="font-weight: bold"/>
                     <p:outputPanel>
-                        <h:dataTable value="#{userBean.currentUser.roles}" var="role"><f:facet name="Role">Status</f:facet>
+                        <h:dataTable value="#{adminBean.currentUser.roles}" var="role"><f:facet name="Role">Status</f:facet>
                             <h:column>
                                 #{role.name}
                             </h:column>
@@ -42,7 +42,7 @@
                 <h:panelGrid style="width:100%" columns="1" cellpadding="80" columnClasses="description, detail" styleClass="commandButton" >
                     <p:growl id="message" showDetail="true" />
 
-                    <p:commandButton value="#{msgs['button.delete']}" action="#{userBean.deleteUser(userBean.currentUser)}" update="message">
+                    <p:commandButton value="#{msgs['button.delete']}" action="#{adminBean.deleteUser(adminBean.currentUser)}" update="message">
                         <p:confirm header="Achtung" message="Möchten Sie dieses Profil wirklich löschen?"/>
                     </p:commandButton>
